+++ package.json
@@ -11,7 +11,7 @@
     "axios": "^1.3.4",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-router-dom": "^6.8.1",
+    "react-router-dom": "^5.3.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
