+++ src/index.js
@@ -1,6 +1,5 @@
 
 // react 17.0.2
-
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
@@ -9,6 +8,7 @@ import App from './App';
 ReactDOM.render(
   <React.StrictMode>
     <App/>
-  </React.StrictMode>,
+  </React.StrictMode>
+	,
   document.getElementById('root')
 );
